Sync from DMOJ 2017

+++ dmoj/cptbox/helper.cpp
@@ -15,6 +15,9 @@
 #   include <sys/socket.h>
 #   include <sys/sysctl.h>
 #   include <libprocstat.h>
+#else
+// No ASLR on FreeBSD... not as of 11.0, anyway
+#   include <sys/personality.h>
 #endif
 
 #if defined(__FreeBSD__) || (defined(__APPLE__) && defined(__MACH__))
@@ -22,6 +25,36 @@
 #else
 #   define FD_DIR "/proc/self/fd"
 #endif
+
+pt_debugger *get_ptdebugger(int type) {
+    switch (type) {
+#ifdef HAS_DEBUGGER_X86
+        case DEBUGGER_X86:
+            return new pt_debugger_x86();
+#endif
+#ifdef HAS_DEBUGGER_X64
+        case DEBUGGER_X64:
+            return new pt_debugger_x64();
+#endif
+#ifdef HAS_DEBUGGER_X86_ON_X64
+        case DEBUGGER_X86_ON_X64:
+            return new pt_debugger_x86_on_x64();
+#endif
+#ifdef HAS_DEBUGGER_X32
+        case DEBUGGER_X32:
+            return new pt_debugger_x32();
+#endif
+#ifdef HAS_DEBUGGER_ARM
+        case DEBUGGER_ARM:
+            return new pt_debugger_arm();
+#endif
+#ifdef HAS_DEBUGGER_ARM64
+        case DEBUGGER_ARM64:
+            return new pt_debugger_arm64();
+#endif
+    }
+    return NULL;
+}
 
 inline void setrlimit2(int resource, rlim_t cur, rlim_t max) {
     rlimit limit;
@@ -35,6 +68,12 @@ inline void setrlimit2(int resource, rlim_t limit) {
 }
 
 int cptbox_child_run(const struct child_config *config) {
+#ifndef __FreeBSD__
+    // There is no ASLR on FreeBSD, but disable it elsewhere
+    if (config->personality > 0)
+        personality(config->personality);
+#endif
+
     if (config->address_space)
         setrlimit2(RLIMIT_AS, config->address_space);
 
@@ -53,19 +92,20 @@ int cptbox_child_run(const struct child_config *config) {
     setrlimit2(RLIMIT_STACK, RLIM_INFINITY);
     setrlimit2(RLIMIT_CORE, 0);
 
-    if (config->stdin >= 0)  dup2(config->stdin, 0);
-    if (config->stdout >= 0) dup2(config->stdout, 1);
-    if (config->stderr >= 0) dup2(config->stderr, 2);
+    if (config->stdin_ >= 0)  dup2(config->stdin_, 0);
+    if (config->stdout_ >= 0) dup2(config->stdout_, 1);
+    if (config->stderr_ >= 0) dup2(config->stderr_, 2);
 
     for (int i = 3; i <= config->max_fd; ++i)
         dup2(config->fds[i-3], i);
 
     cptbox_closefrom(config->max_fd + 1);
 
-    ptrace_traceme();
+    if (ptrace_traceme())
+        return 204;
     kill(getpid(), SIGSTOP);
     execve(config->file, config->argv, config->envp);
-    return 3306;
+    return 205;
 }
 
 // From python's _posixsubprocess
